+++ cmd/cortextool/pkg/commands/rules.go
@@ -117,6 +117,22 @@ func (r *RuleCommand) Register(app *kingpin.Application) {
 			Envar("CORTEX_TENANT_ID").
 			Required().
 			StringVar(&r.ClientConfig.ID)
+
+		c.Flag("tls-ca-path", "TLS CA certificate to verify cortex API as part of mTLS, alternatively set CORTEX_TLS_CA_PATH.").
+			Default("").
+			Envar("CORTEX_TLS_CA_CERT").
+			StringVar(&r.ClientConfig.TLS.CAPath)
+
+		c.Flag("tls-cert-path", "TLS client certificate to authenticate with cortex API as part of mTLS, alternatively set CORTEX_TLS_CERT_PATH.").
+			Default("").
+			Envar("CORTEX_TLS_CLIENT_CERT").
+			StringVar(&r.ClientConfig.TLS.CertPath)
+
+		c.Flag("tls-key-path", "TLS client certificate private key to authenticate with cortex API as part of mTLS, alternatively set CORTEX_TLS_KEY_PATH.").
+			Default("").
+			Envar("CORTEX_TLS_CLIENT_KEY").
+			StringVar(&r.ClientConfig.TLS.KeyPath)
+
 	}
 
 	// Print Rules Command